fix vite config and build
