+++ frontend/package.json
@@ -2,6 +2,7 @@
   "name": "frontend",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
@@ -43,6 +44,6 @@
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^4.6.0",
-    "vite": "^7.0.0"
+    "vite": "^6.3.5"
   }
 }
