+++ frontend/src/components/OnboardingOverlay.jsx
